requirements.txt
pyserial
matplotlib
ttkthemes
Pillow